--- a/tensordict/utils.py
+++ b/tensordict/utils.py
@@ -18,7 +18,6 @@ import time
 import warnings
 from collections import defaultdict, OrderedDict
 from collections.abc import KeysView
-from concurrent.futures import Future
 from copy import copy
 from functools import wraps
 from importlib import import_module