Test:
```python
import torch
from tensordict import TensorDict
from torch.utils.benchmark import Timer
# torch.set_num_threads(16)
import functools

d = {}
sub = d
for _ in range(20):
    for i in range(50):
        sub[str(i)] = torch.rand(2048, 2048)
    sub["nested"] = {}
    sub = sub["nested"]
td = TensorDict(d, batch_size=[])

print(Timer("td.to('cuda')", globals=globals()).adaptive_autorange())
  Median: 1.02 s
print(Timer("td.pin_memmory().to('cuda')", globals=globals()).adaptive_autorange())
#   Median: 1.80 s
print(Timer("td.to('cuda', pin_memory=True)", globals=globals()).adaptive_autorange())
#   Median: 444.79 ms
print(Timer("td.to('cuda', pin_memory=True, num_threads=0)", globals=globals()).adaptive_autorange())
#   Median: 1.82 s
print(Timer("td.to('cuda', pin_memory=True, num_threads=1)", globals=globals()).adaptive_autorange())
#   Median: 1.82 s
print(Timer("td.to('cuda', pin_memory=True, num_threads=16)", globals=globals()).adaptive_autorange())
#   Median: 375.45 ms
```

cc @albanD @mikaylagawarecki @shagunsodhani @dstaay-fb